# docs/swagger/swagger.yaml
basePath: /api
definitions:
  api.HTTPError:
    properties:
      err:
        type: string
    type: object
...
  model.Chunk:
    properties:
      cars:
        items:
          $ref: '#/definitions/model.Car'
        type: array
      createdAt:
        type: string
      errorMessage:
        type: string
      id:
        type: integer
      itemParts:
        items:
          $ref: '#/definitions/model.ItemPart'
        type: array
      packingState:
        $ref: '#/definitions/model.WorkState'
      packingWorkerId:
        type: string
      sourceId:
        type: integer
    type: object
  model.Dataset:
    properties:
      createdAt:
        type: string
      encryptionRecipients:
        items:
          type: string
        type: array
      encryptionScript:
        type: string
      id:
        type: integer
      maxSize:
        type: integer
      name:
        type: string
      outputDirs:
        items:
          type: string
        type: array
      pieceSize:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.WalletAssignment:
    properties:
      datasetId:
        type: integer
      id:
        type: integer
      walletId:
        type: string
    type: object
...
  store.PieceReader:
    type: object
...
host: localhost:9090
info:
  contact:
    name: Xinan Xu
    url: https://github.com/data-preservation-programs/singularity/issues
  description: This is the API for Singularity, a tool for large-scale clients with
    PB-scale data onboarding to Filecoin network.
  license:
    name: MIT + Apache 2.0
    url: https://github.com/data-preservation-programs/singularity/blob/main/LICENSE
  title: Singularity API
  version: beta
paths:
  /admin/init:
    post:
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Initialize the database
      tags:
      - Admin
  /admin/reset:
    post:
      description: This will drop all tables and recreate them.
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Reset the database
      tags:
      - Admin
  /chunk/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Chunk ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Chunk'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get detail of a specific chunk
      tags:
      - Data Source
  /dataset:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Dataset'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: List all datasets
      tags:
      - Dataset
    post:
      consumes:
      - application/json
      description: The dataset is a top level object to distinguish different dataset.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dataset.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Dataset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Create a new dataset
      tags:
      - Dataset
  /dataset/{datasetName}:
    delete:
      description: Important! If the dataset is large, this command will take some
        time to remove all relevant data.
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Remove a specific dataset. This will not remove the CAR files.
      tags:
      - Dataset
    patch:
      consumes:
      - application/json
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dataset.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Dataset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update a dataset
      tags:
      - Dataset
  /dataset/{datasetName}/piece:
    get:
      consumes:
      - application/json
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Car'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: List all pieces for the dataset that are available for deal making
      tags:
      - Dataset
    post:
      consumes:
      - application/json
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dataset.AddPieceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Car'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Manually register a piece (CAR file) with the dataset for deal making
        purpose
      tags:
      - Dataset
  /dataset/{datasetName}/wallet:
    get:
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Wallet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: List all wallets of a dataset.
      tags:
      - Wallet
  /dataset/{datasetName}/wallet/{wallet}:
    delete:
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Wallet Address
        in: path
        name: wallet
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Remove an associated wallet from a dataset
      tags:
      - Wallet
    post:
      consumes:
      - application/json
      parameters:
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Wallet Address
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WalletAssignment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Associate a new wallet with a dataset
      tags:
      - Wallet Association
  /deal:
    post:
      consumes:
      - application/json
      description: List all deals
      parameters:
      - description: ListDealRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/deal.ListDealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Deal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: List all deals
      tags:
      - Deal
  /item/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Item'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get details about an item
      tags:
      - Data Source
  /piece/{id}/metadata:
    get:
      description: Get metadata for a piece for how it may be reassembled from the
        data source
      parameters:
      - description: Piece CID
        in: path
        name: piece
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.PieceReader'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get metadata for a piece
      tags:
      - Metadata
  /schedule:
    post:
      consumes:
      - application/json
      description: Create a new schedule
      parameters:
      - description: CreateRequest
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/schedule.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Create a new schedule
      tags:
      - Deal Schedule
  /schedule/{id}/pause:
    post:
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Pause a specific schedule
      tags:
      - Deal Schedule
  /schedule/{id}/resume:
    post:
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Resume a specific schedule
      tags:
      - Deal Schedule
...
  /send_deal:
    post:
      consumes:
      - application/json
      description: Send a manual deal proposal
      parameters:
      - description: Proposal
        in: body
        name: proposal
        required: true
        schema:
          $ref: '#/definitions/deal.Proposal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Deal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Send a manual deal proposal
      tags:
      - Deal
  /source:
    get:
      parameters:
      - description: Dataset name
        in: query
        name: dataset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Source'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: List all sources for a dataset
      tags:
      - Data Source
  /source/{id}:
    delete:
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Remove a source
      tags:
      - Data Source
    patch:
      consumes:
      - application/json
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      - description: Config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/datasource.AllConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Source'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update the config options of a source
      tags:
      - Data Source
  /source/{id}/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datasource.CheckSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_data-preservation-programs_singularity_handler_datasource.Entry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Check the connection of the data source by listing a path
      tags:
      - Data Source
  /source/{id}/chunks:
    get:
      consumes:
      - application/json
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Car'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get all dag details of a data source
      tags:
      - Data Source
  /source/{id}/daggen:
    post:
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Source'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Mark a source as ready for DAG generation
      tags:
      - Data Source
  /source/{id}/items:
    get:
      consumes:
      - application/json
      parameters:
      - description: Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Item'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get all item details of a data source
      tags:
      - Data Source
...
swagger: "2.0"